Uploaded car.jl
I demonstrate object oriented programming in Julia using my car blueprint as in some of my other projects. Julia is often times considered not to be a true object oriented language, as it does support composite types in the form of structs, but generally does not support member functions or methods within the structs, other than the constructor. For this reason, I managed to use object oriented programming with C in a way that better suits OOP principles than I could with Julia. Other than Clojure, I would say Julia is the least OOP language in this series of ObjectOriented[Lang].